+++ ChangeLog.md
@@ -189,6 +189,7 @@ ChangeLog
 * Fixed: Restoring old setting after changing libxml_disable_entity_loader.
 * Fixed: Issue #422 Preconditions were not being set on PUT on non-existant
   files. Not really a chance for data-loss, but incorrect nevertheless.
+* Fixed: Issue #427: Now checking preconditions on DELETE requests.
 
 
 1.7.11 (2014-02-26)